refactor: dedup shared helpers across services and cli
- share resolveChange (change-plan/change-tasks) via services/change-resolver;
  read metadata.yaml once per run instead of twice
- share resolveLogLevel across all 10 commands via cli/log-level
