+++ src/cli/commands/knowledge-init.ts
@@ -3,17 +3,8 @@ import { execute } from '../../services/knowledge-init.service.js';
 import { formatKnowledgeInitOutput } from '../formatters/knowledge-init-output.js';
 import { handleError } from '../formatters/error-output.js';
 import type { GlobalOptions } from '../index.js';
-import type { LogLevel } from '../../types/config.js';
 import { parseDepth } from '../parse-options.js';
-
-/**
- * Resolve log level from global options.
- */
-function resolveLogLevel(opts: GlobalOptions): LogLevel {
-  if (opts.quiet) return 'quiet';
-  if (opts.verbose) return 'verbose';
-  return 'normal';
-}
+import { resolveLogLevel } from '../log-level.js';
 
 /**
  * Register the `init` subcommand under the `knowledge` command group.
