+++ src/cli/commands/measure.ts
@@ -3,14 +3,8 @@ import { execute } from '../../services/measure.service.js';
 import { formatMeasureOutput } from '../formatters/measure-output.js';
 import { handleError } from '../formatters/error-output.js';
 import type { GlobalOptions } from '../index.js';
-import type { LogLevel } from '../../types/config.js';
 import { DEFAULT_REPORT_FILENAME } from '../../types/measurement.js';
-
-function resolveLogLevel(opts: GlobalOptions): LogLevel {
-  if (opts.quiet) return 'quiet';
-  if (opts.verbose) return 'verbose';
-  return 'normal';
-}
+import { resolveLogLevel } from '../log-level.js';
 
 /**
  * Register the `measure` command.
