+++ src/services/change-plan.service.ts
@@ -1,12 +1,12 @@
 import * as fs from 'node:fs';
 import * as path from 'node:path';
-import { select } from '@inquirer/prompts';
 import { PrerequisiteError } from '../types/errors.js';
 import { readConfig } from '../lib/config.js';
 import { atomicWrite } from '../lib/fs-utils.js';
 import { renderTemplate } from '../lib/template.js';
 import { parseYaml, stringifyYaml } from '../lib/yaml-utils.js';
 import type { ChangeMetadata } from '../types/change.js';
+import { resolveChange } from './change-resolver.js';
 
 export interface ChangePlanOptions {
   change?: string;
@@ -36,7 +36,12 @@ export async function execute(options: ChangePlanOptions): Promise<ChangePlanRes
   await readConfig(cwd);
 
   // 2. Resolve change name
-  const changeName = await resolveChange(cwd, options.change, options.quiet);
+  const changeName = await resolveChange(
+    cwd,
+    options.change,
+    options.quiet,
+    'Select the change to generate a plan for:',
+  );
 
   const changeDir = path.join(cwd, '.prospec', 'changes', changeName);
 
@@ -49,14 +54,12 @@ export async function execute(options: ChangePlanOptions): Promise<ChangePlanRes
     );
   }
 
-  // 4. Read metadata for related_modules
+  // 4. Read metadata ONCE — reused for related_modules and the status update
   const metadataPath = path.join(changeDir, 'metadata.yaml');
-  let relatedModules: string[] = [];
-  if (fs.existsSync(metadataPath)) {
-    const metadataContent = fs.readFileSync(metadataPath, 'utf-8');
-    const metadata = parseYaml<ChangeMetadata>(metadataContent, metadataPath);
-    relatedModules = metadata.related_modules ?? [];
-  }
+  const metadata = fs.existsSync(metadataPath)
+    ? parseYaml<ChangeMetadata>(fs.readFileSync(metadataPath, 'utf-8'), metadataPath)
+    : null;
+  const relatedModules = metadata?.related_modules ?? [];
 
   // 5. Build template context
   const templateContext = {
@@ -80,8 +83,11 @@ export async function execute(options: ChangePlanOptions): Promise<ChangePlanRes
   await atomicWrite(deltaSpecPath, deltaSpecContent);
   createdFiles.push(`.prospec/changes/${changeName}/delta-spec.md`);
 
-  // 8. Update metadata.yaml status to 'plan'
-  await updateMetadataStatus(metadataPath);
+  // 8. Update metadata.yaml status to 'plan' (write the already-parsed object)
+  if (metadata) {
+    metadata.status = 'plan';
+    await atomicWrite(metadataPath, stringifyYaml(metadata));
+  }
 
   return {
     changeName,
@@ -89,88 +95,4 @@ export async function execute(options: ChangePlanOptions): Promise<ChangePlanRes
     createdFiles,
     relatedModules,
   };
-}
-
-/**
- * Resolve which change to work on.
- *
- * Strategy (per contracts/cli-commands.md):
- * 1. If --change is provided → use it directly
- * 2. Scan .prospec/changes/ directory
- * 3. If 1 change → auto-select
- * 4. If multiple → interactive prompt (or error in --quiet mode)
- * 5. If 0 changes → PrerequisiteError
- */
-async function resolveChange(
-  cwd: string,
-  explicitChange: string | undefined,
-  quiet: boolean | undefined,
-): Promise<string> {
-  if (explicitChange) {
-    // Validate the specified change exists
-    const changeDir = path.join(cwd, '.prospec', 'changes', explicitChange);
-    if (!fs.existsSync(changeDir)) {
-      throw new PrerequisiteError(
-        `Change '${explicitChange}' not found`,
-        'Verify the change name is correct, or run `prospec change story` to create a new change',
-      );
-    }
-    return explicitChange;
-  }
-
-  const changesDir = path.join(cwd, '.prospec', 'changes');
-  if (!fs.existsSync(changesDir)) {
-    throw new PrerequisiteError(
-      'No changes found',
-      'Run `prospec change story <name>` first to create a change request',
-    );
-  }
-
-  const entries = fs.readdirSync(changesDir, { withFileTypes: true });
-  const changeNames = entries
-    .filter((e) => e.isDirectory())
-    .map((e) => e.name);
-
-  if (changeNames.length === 0) {
-    throw new PrerequisiteError(
-      'No changes found',
-      'Run `prospec change story <name>` first to create a change request',
-    );
-  }
-
-  if (changeNames.length === 1) {
-    return changeNames[0]!;
-  }
-
-  // Multiple changes found
-  if (quiet) {
-    throw new PrerequisiteError(
-      `Multiple changes found: ${changeNames.join(', ')}`,
-      'Use `--change <name>` to specify which change to use',
-    );
-  }
-
-  // Interactive prompt
-  const selected = await select({
-    message: 'Select the change to generate a plan for:',
-    choices: changeNames.map((name) => ({ name, value: name })),
-  });
-
-  return selected;
-}
-
-/**
- * Update metadata.yaml status to 'plan'.
- * Preserves existing fields and comments where possible.
- */
-async function updateMetadataStatus(
-  metadataPath: string,
-): Promise<void> {
-  if (!fs.existsSync(metadataPath)) return;
-
-  const content = fs.readFileSync(metadataPath, 'utf-8');
-  const metadata = parseYaml<ChangeMetadata>(content, metadataPath);
-  metadata.status = 'plan';
-  const updated = stringifyYaml(metadata);
-  await atomicWrite(metadataPath, updated);
 }
